+++ mobile/android/android-components/docs/changelog.md
@@ -19,6 +19,7 @@ permalink: /changelog/
 
 * **service-glean**
   * Hyphens `-` are now allowed in labels for metrics.  See [1566764](https://bugzilla.mozilla.org/show_bug.cgi?id=1566764).
+  * ⚠️ **This is a breaking change**: Timespan values are returned in their configured time unit in the testing API.
 
 * **support-ktx**
   * ⚠️ **This is a breaking behavior change**: `JSONArray.mapNotNull` is now an inline function, changing the behavior of the `return` keyword within its lambda.